+++ src/main/java/ProductCatalog/servlets/ProductServlet.java
@@ -1,12 +1,12 @@
 package ProductCatalog.servlets;
 
 import ProductCatalog.annotations.Auditable;
+import ProductCatalog.constants.Permission;
 import ProductCatalog.dto.ProductDTO;
 import ProductCatalog.mappers.ProductMapper;
 import ProductCatalog.models.Product;
-import ProductCatalog.models.Role;
-import ProductCatalog.models.User;
 import ProductCatalog.services.implemetations.ProductService;
+import ProductCatalog.utils.AccessUtil;
 import ProductCatalog.validators.ProductValidator;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
@@ -14,7 +14,6 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
 
 import java.io.IOException;
 import java.util.List;
@@ -32,7 +31,6 @@ public class ProductServlet extends HttpServlet {
     @Auditable(action = "get all products")
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
         String catalogIdStr = req.getParameter("catalogId");
 
         List<Product> products =
@@ -51,14 +49,7 @@ public class ProductServlet extends HttpServlet {
     @Auditable(action = "create new product")
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.setContentType("application/json");
-
-        HttpSession session = req.getSession();
-        Object user = session.getAttribute(SessionAttributes.USER);
-        if (!((User) user).getRole().equals(Role.ADMIN)) {
-            resp.setStatus(HttpServletResponse.SC_FORBIDDEN);
-            mapper.writeValue(resp.getWriter(), "You are not allowed to perform this action");
-        }
+        AccessUtil.checkPermission(req, Permission.CREATE_PRODUCT);
 
         ProductDTO dto = mapper.readValue(req.getInputStream(), ProductDTO.class);
 
@@ -76,6 +67,7 @@ public class ProductServlet extends HttpServlet {
 
         boolean created = productService.createProduct(product);
 
+        resp.setContentType("application/json");
         if (created) {
             resp.setStatus(HttpServletResponse.SC_CREATED);
             mapper.writeValue(resp.getWriter(), ProductMapper.INSTANCE.toDTO(product));
@@ -89,13 +81,9 @@ public class ProductServlet extends HttpServlet {
     @Auditable(action = "delete product")
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        AccessUtil.checkPermission(req, Permission.DELETE_PRODUCT);
+
         long id = Long.parseLong(req.getParameter("id"));
-        HttpSession session = req.getSession();
-        Object user = session.getAttribute(SessionAttributes.USER);
-        if (!((User) user).getRole().equals(Role.ADMIN)) {
-            resp.setStatus(HttpServletResponse.SC_FORBIDDEN);
-            mapper.writeValue(resp.getWriter(), "You are not allowed to perform this action");
-        }
         boolean deleted = productService.deleteProduct(id);
         if (deleted) {
             resp.setStatus(HttpServletResponse.SC_OK);
